Bugs en verbeteringen in het toevoegen van dranken

+++ NewDirectory/MID2/MID/ViewController.h
@@ -79,7 +79,7 @@
     float *perfectint;
     
     
-    int preciesglazen;
+    float preciesglazen;
     
     
     int een;
